+++ Modules/Intent.Modules.Eventing.MassTransit.Scheduling/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.2
+
+- Improvement: Updates based on changes made in `Intent.Eventing.MassTransit`.
+
 ### Version 1.0.1
 
 - Fixed: Missing using directives.